-Converted CJS modules to ES Modules. -Added eslint plugin compat to help with linting JS in VSCode. -Added esbuild to handle bundling the source files. -Updated npm dependencies to their latest versions.
