+++ .editorconfig
@@ -2,8 +2,8 @@ root = true
 
 [*]
 end_of_line = lf
-insert_final_newline = true
+insert_final_newline = false
 trim_trailing_whitespace = true
 charset = utf-8
-indent_style = space
+indent_style = tab
 indent_size = 2
