+++ .eslintrc.json
@@ -1,7 +1,16 @@
 {
-  "root": true,
-  "extends": "@domenic",
-  "env": {
-    "node": true
-  }
+	"env": {
+		"browser": true,
+		"node": true,
+		"es2022": true,
+		"jest": true
+	},
+	"plugins": ["compat"],
+	"extends": ["eslint:recommended", "plugin:compat/recommended"],
+	"parserOptions": {
+		"ecmaVersion": "latest",
+		"sourceType": "module"
+	},
+	"rules": {},
+	"overrides": []
 }
