+++ .vscode/settings.json
@@ -1,0 +1,9 @@
+{
+	"cSpell.words": [
+		"compat",
+		"esbuild",
+		"outfile",
+		"Subscribr",
+		"WHATWG"
+	]
+}
