+++ README.md
@@ -1,9 +1,10 @@
 # Parse, serialize, and manipulate MIME types
 
 This package will parse [MIME types](https://mimesniff.spec.whatwg.org/#understanding-mime-types) into a structured format, which can then be manipulated and serialized:
+This version is using ES Modules instead of commonJS.
 
 ```js
-const MIMEType = require("whatwg-mimetype");
+import MIMEType from '@d1g1tal/mime-type';
 
 const mimeType = new MIMEType(`Text/HTML;Charset="utf-8"`);
 
@@ -92,8 +93,8 @@ mimeType.parameters.set("@", "x");
 If you want primitives on which to build your own API, you can get direct access to the parsing and serialization algorithms as follows:
 
 ```js
-const parse = require("whatwg-mimetype/parser");
-const serialize = require("whatwg-mimetype/serialize");
+import parse from '@d1g1tal/mime-type/parser.js';
+import serialize from '@d1g1tal/mime-type/serializer.js';
 ```
 
 `parse(string)` returns an object containing the `type` and `subtype` strings, plus `parameters`, which is a `Map`. This is roughly our equivalent of the spec's [MIME type record](https://mimesniff.spec.whatwg.org/#mime-type). If parsing fails, it instead returns `null`.
