+++ jsconfig.json
@@ -1,0 +1,12 @@
+	{
+		"compilerOptions": {
+			"baseUrl": "./",
+			"checkJs": false,
+			"module": "ES2022",
+			"target": "ES2022",
+			"noImplicitAny": false,
+			"moduleResolution": "NodeNext"
+		},
+		"include": [ "./*", "src/**/*", "tests/*.js" ],
+		"exclude": [ "node_modules", "**/node_modules/*", "dist" ]
+	}
