+++ package.json
@@ -1,37 +1,59 @@
 {
-  "name": "whatwg-mimetype",
+  "name": "@d1g1tal/mime-type",
   "description": "Parses, serializes, and manipulates MIME types, according to the WHATWG MIME Sniffing Standard",
-  "keywords": [
+  "version": "3.0.1",
+  "author": "Domenic Denicola <d@domenic.me> (https://domenic.me/)",
+	"maintainers": [{
+		"name": "Jason DiMeo",
+    "email": "jason.dimeo@gmail.com"
+	}],
+  "license": "MIT",
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/D1g1talEntr0py/mime-type.git"
+  },
+	"publishConfig": {
+    "access": "public"
+  },
+	"type": "module",
+	"module": "./dist/esm/mime-type.min.js",
+  "files": [
+    "src/"
+  ],
+	"keywords": [
     "content-type",
     "mime type",
     "mimesniff",
     "http",
     "whatwg"
   ],
-  "version": "3.0.0",
-  "author": "Domenic Denicola <d@domenic.me> (https://domenic.me/)",
-  "license": "MIT",
-  "repository": "jsdom/whatwg-mimetype",
-  "main": "lib/mime-type.js",
-  "files": [
-    "lib/"
-  ],
   "scripts": {
-    "test": "jest",
-    "coverage": "jest --coverage",
-    "lint": "eslint .",
+		"build": "rimraf dist/esm && esbuild src/mime-type.js --bundle --target=es2022 --format=esm --outfile=dist/esm/mime-type.js && esbuild src/mime-type.js --bundle --sourcemap --minify --target=es2020 --format=esm --outfile=dist/esm/mime-type.min.js",
+		"build:browser": "rimraf dist/browser && esbuild src/mime-type.js --bundle --platform=browser --global-name=MIMEType --footer:js='window.MIMEType = MIMEType.default;' --outfile=dist/browser/mime-type.js && esbuild src/mime-type.js --bundle --sourcemap --minify --platform=browser --global-name=MIMEType --footer:js='window.MIMEType = MIMEType.default;' --outfile=dist/browser/mime-type.min.js",
+    "test": "node --no-warnings --experimental-vm-modules node_modules/jest/bin/jest.js",
+    "coverage": "node --no-warnings --experimental-vm-modules node_modules/jest/bin/jest.js --coverage",
     "pretest": "node scripts/get-latest-platform-tests.js"
   },
   "devDependencies": {
-    "@domenic/eslint-config": "^1.4.0",
-    "eslint": "^7.32.0",
-    "jest": "^27.2.0",
-    "minipass-fetch": "^1.4.1",
+    "esbuild": "^0.15.10",
+    "eslint": "^8.24.0",
+    "eslint-plugin-compat": "^4.0.2",
+    "jest": "^29.1.2",
+    "minipass-fetch": "^2.1.2",
     "printable-string": "^0.3.0",
+    "rimraf": "^3.0.2",
     "whatwg-encoding": "^2.0.0"
   },
+  "browserslist": [
+    ">= 0.5%",
+    "last 2 major versions",
+    "Firefox ESR",
+    "not dead",
+    "not ios_saf < 15",
+    "not op_mini all"
+  ],
   "engines": {
-    "node": ">=12"
+    "node": ">=16"
   },
   "jest": {
     "coverageDirectory": "coverage",
