+++ src/utils.js
@@ -1,31 +1,29 @@
-"use strict";
-
-exports.removeLeadingAndTrailingHTTPWhitespace = string => {
+export function removeLeadingAndTrailingHTTPWhitespace(string) {
   return string.replace(/^[ \t\n\r]+/u, "").replace(/[ \t\n\r]+$/u, "");
-};
+}
 
-exports.removeTrailingHTTPWhitespace = string => {
+export function removeTrailingHTTPWhitespace(string) {
   return string.replace(/[ \t\n\r]+$/u, "");
-};
+}
 
-exports.isHTTPWhitespaceChar = char => {
+export function isHTTPWhitespaceChar(char) {
   return char === " " || char === "\t" || char === "\n" || char === "\r";
-};
+}
 
-exports.solelyContainsHTTPTokenCodePoints = string => {
+export function solelyContainsHTTPTokenCodePoints(string) {
   return /^[-!#$%&'*+.^_`|~A-Za-z0-9]*$/u.test(string);
-};
+}
 
-exports.soleyContainsHTTPQuotedStringTokenCodePoints = string => {
+export function solelyContainsHTTPQuotedStringTokenCodePoints(string) {
   return /^[\t\u0020-\u007E\u0080-\u00FF]*$/u.test(string);
-};
+}
 
-exports.asciiLowercase = string => {
+export function asciiLowercase(string) {
   return string.replace(/[A-Z]/ug, l => l.toLowerCase());
-};
+}
 
 // This variant only implements it with the extract-value flag set.
-exports.collectAnHTTPQuotedString = (input, position) => {
+export function collectAnHTTPQuotedString(input, position) {
   let value = "";
 
   position++;
@@ -57,4 +55,4 @@ exports.collectAnHTTPQuotedString = (input, position) => {
   }
 
   return [value, position];
-};
+}
